requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.2.10           # via django
astroid==2.4.2            # via pylint
autopep8==1.5.3           # via -r requirements.in
django==3.0.7             # via -r requirements.in, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
factory_boy==2.12.0       # via -r requirements.in
faker==4.1.1              # via factory-boy
flake8==3.8.3             # via -r requirements.in
isort==4.3.21             # via pylint
lazy-object-proxy==1.4.3  # via astroid
mccabe==0.6.1             # via flake8, pylint
pycodestyle==2.6.0        # via autopep8, flake8
pyflakes==2.2.0           # via flake8
pylint-django==2.0.15     # via -r requirements.in
pylint-plugin-utils==0.6  # via pylint-django
pylint==2.5.3             # via pylint-django, pylint-plugin-utils
python-dateutil==2.8.1    # via faker
pytz==2020.1              # via django
six==1.15.0               # via astroid, python-dateutil
sqlparse==0.3.1           # via django
text-unidecode==1.3       # via faker
toml==0.10.1              # via autopep8, pylint
wrapt==1.12.1             # via astroid
